requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.6
pdfplumber>=0.10.3
python-docx>=1.1.0
openai>=1.54.0
faiss-cpu>=1.13.1
numpy>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
